+++ src/App.css
@@ -14,7 +14,6 @@ body {
 }
 
 
-
 .selector-button
 {
 	font-size: 15px;
@@ -309,7 +308,8 @@ overflow-wrap: break-word;
 /* NavHeader */
 
 header {
-	background-color: red;
+	background-color: white;
+	border-bottom: 2px solid darkblue;
 	overflow: hidden;
 	display: flex;
 	padding: 1px 1px;
@@ -319,14 +319,24 @@ nav{
 	height: 5vh;
 }
 nav .login-button {
-    font-size: 16px;
 	position:absolute;
-	right:0;
+	right:5px;
 	top: 5px;
-	margin-right: 5px;
-
 }
 
+nav .nav-addFood-btn{
+	position:absolute;
+	left: 5px;
+	top: 5px;
+}
+nav button {
+	font-size: 16px;
+	margin-right: 5px;
+	background-color: rgb(103, 103, 201);
+	color: white;
+	border: 2px solid darkblue;
+	border-radius:6px;
+}
 /* Login Page */
 .login-modal{
 	position: absolute;
@@ -336,11 +346,11 @@ nav .login-button {
     top: 50px;
     margin-left: auto;
     margin-right: auto;
-    max-width: 360px;
-    height: 300px;
+	max-width: 30vw;
+    height: 25vh;
     background-color: black;
     border-radius: 10px;
-    padding: 40px;
+    padding: 60px;
 }
 .login-modal input{
 	color: black;
@@ -351,4 +361,36 @@ nav .login-button {
     right: 5px;
 	top: 4px;
 	background-color:black
+}
+
+.login-btn{
+	color: black;
+}
+
+/* Add New Food page */
+
+.newFood-modal{
+	position: absolute;
+	color:aliceblue;
+    left: 5vw;
+    top: 3vh;
+    margin-left: auto;
+    margin-right: auto;
+    max-width: 20vw;
+    height: 15vh;
+    background-color: darkblue;
+	opacity: 90%;
+    border-radius: 10px;
+    padding: 40px;
+}
+
+
+/* Home page */
+
+.HomePAGE{
+	background-image: url(https://wallpapercave.com/wp/wp8217848.jpg);
+	background-repeat: no-repeat;
+  	background-size: cover;
+	height:100vh;
+	width:100vw;
 }
